Adjusting PriceQuery to only return the applicable price according to filters. New test for searching an existing price

+++ src/main/java/com/inditex/priceservice/domain/model/Price.java
@@ -4,7 +4,7 @@ import java.math.BigDecimal;
 import java.time.Instant;
 
 public record Price(
-        Long brandId,
+        Integer brandId,
         Instant startDate,
         Instant endDate,
         Long priceList,
